[do not merge] update pop merkle tree (#276)
* update pop merkle tree

* fix

* fix

* With newest altintegration

* Fix build

* Add version to encoded popdata

* Fix merkle code

+++ src/bootstraps.h
@@ -35,7 +35,7 @@ struct AltChainParamsVBTC : public altintegration::AltChainParams {
         return bootstrap;
     }
 
-    uint32_t getIdentifier() const noexcept override
+    int64_t getIdentifier() const noexcept override
     {
         return 0x3ae6ca;
     }
